Correción del error en la división
He casteado los valores int pasados al método división antes de hacer la
operación con ellos, si no no devuelve los decimales.

+++ tarea5/src/tarea5/Ejerc13App.java
@@ -82,7 +82,7 @@ public class Ejerc13App {
 			JOptionPane.showMessageDialog(null, "No se puede dividir entre 0!");
 			System.exit(0);
 		}	
-		return (a / b);
+		return ((double)a / (double)b);
 	}
 	
 	public static int potencia (int a, int b) {
